Bilibili/requirements.txt
pandas==0.25.2
tensorflow-serving-api==1.14.0